Pagamento com cartão de crédito
O sistema de pagamento com o cartão de crédito está funcionando.

+++ mysql/banco_de_dados_mysql.sql
@@ -24,7 +24,7 @@ create table cadastro_do_cliente (
     id int(5) primary key not null auto_increment,
     nome text(30) not null,
     email varchar(30) not null unique,
-    cpf int(5) not null unique,
+    cpf bigint(14) not null unique,
     senha varchar(30) not null,
     ddd_telefone int(3) not null,
     telefone int(9),
